full python in client

+++ python/python/psyche/models/hf_transformers.py
@@ -78,7 +78,7 @@ class HfTransformersAuto(CausalLM):
         tp: int = 1,
         override_max_position_embeddings: Optional[int] = None,
         param_dtype: torch.dtype = torch.bfloat16,
-        reduce_dtype: torch.dtype = torch.float32,
+        reduce_dtype: torch.dtype = torch.bfloat16,
         fsdp_modules: Optional[Iterable[str]] = None,
     ):
         if isinstance(source, PretrainedSourceStateDict):
@@ -134,7 +134,7 @@ class HfTransformersAuto(CausalLM):
                     param_dtype=param_dtype, reduce_dtype=reduce_dtype
                 )
                 fsdp_config = {
-                    #"mesh": world_mesh[tuple(("dp_replicate",))],
+                    # "mesh": world_mesh[tuple(("dp_replicate",))],
                     "mesh": world_mesh[tuple(("dp_shard",))],
                     "mp_policy": mp_policy,
                 }
